pyproject.toml
[project]
name = "app-fastapi-sqlmodel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi>=0.116.1",
    "passlib[bcrypt]==1.7.4",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]

[tool.setuptools]
package-dir = {"" = "src"} 

[tool.setuptools.packages.find]
where = ["src"]
